model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
pillow>=9.5.0
opencv-python>=4.5.0,<4.10.0
webdataset>=0.2.48
pyyaml>=6.0
boto3>=1.28.0
awscli>=1.29.0
tqdm>=4.65.0
tensorboard>=2.13.0
wandb>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
ftfy>=6.1.1
regex>=2023.6.3
git+https://github.com/openai/CLIP.git

